online_inference/requirements.txt
click>7.0.0
fastapi==0.65.1
uvicorn==0.13.4
joblib==1.0.1
pydantic==1.8.2
pandas==1.2.4
marshmallow-dataclass==8.4.1
pyyaml==5.4.1
numpy==1.20.3
requests==2.25.1
scikit-learn==0.24.2
typeguard==2.12.0